football-predictor/telegram-bot/requirements.txt
# Telegram Bot Dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
asyncio==3.4.3

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Background tasks
apscheduler==3.10.4

# Logging
loguru==0.7.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1